requirements.txt
flask
werkzeug
mysql-connector-python
openai
python-dotenv
pandas
rapidfuzz
pillow
